fix: Update UserApiLimit model to fix bug in creating new userApiLimit
This commit updates the UserApiLimit model in the Prisma schema to fix a bug that occurred while creating a new userApiLimit. The bug caused the console.log statements to be executed even when creating a new userApiLimit was successful. This update removes the unnecessary console.log statements and improves the code logic for creating new userApiLimits.