feat: Add error toast for 500 status response in video page
This commit adds an error toast notification in the video page when the server responds with a 500 status code. The error message from the server response is displayed to the user using the toast.error() function. This improves the user experience by providing feedback when an error occurs during video submission.